POC-2AMS-1-main/requirements.txt
pandas>=2.0.0
weaviate-client>=4.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pyyaml>=6.0
boto3>=1.34.0